requirements.txt
lxml_html_clean
torch
transformers
djangorestframework
newspaper4k